break up monolithic descriptor.go into three files
