+++ desc/descriptor.go
@@ -2,20 +2,16 @@ package desc
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
-	"reflect"
 	"sort"
 	"strconv"
 	"strings"
-	"sync"
 	"unicode/utf8"
 
 	"github.com/golang/protobuf/proto"
 	dpb "github.com/golang/protobuf/protoc-gen-go/descriptor"
 
-	descinternal "github.com/jhump/protoreflect/desc/internal"
-	"github.com/jhump/protoreflect/internal"
+	"github.com/jhump/protoreflect/desc/internal"
 )
 
 // Descriptor is the common interface implemented by all descriptor objects.
@@ -48,7 +44,7 @@ type Descriptor interface {
 	AsProto() proto.Message
 }
 
-type sourceInfoRecomputeFunc = descinternal.SourceInfoComputeFunc
+type sourceInfoRecomputeFunc = internal.SourceInfoComputeFunc
 
 // FileDescriptor describes a proto source file.
 type FileDescriptor struct {
@@ -63,194 +59,12 @@ type FileDescriptor struct {
 	services   []*ServiceDescriptor
 	fieldIndex map[string]map[int32]*FieldDescriptor
 	isProto3   bool
-	sourceInfo descinternal.SourceInfoMap
+	sourceInfo internal.SourceInfoMap
 	sourceInfoRecomputeFunc
 }
 
-// CreateFileDescriptor instantiates a new file descriptor for the given descriptor proto.
-// The file's direct dependencies must be provided. If the given dependencies do not include
-// all of the file's dependencies or if the contents of the descriptors are internally
-// inconsistent (e.g. contain unresolvable symbols) then an error is returned.
-func CreateFileDescriptor(fd *dpb.FileDescriptorProto, deps ...*FileDescriptor) (*FileDescriptor, error) {
-	ret := &FileDescriptor{
-		proto:      fd,
-		symbols:    map[string]Descriptor{},
-		fieldIndex: map[string]map[int32]*FieldDescriptor{},
-	}
-	pkg := fd.GetPackage()
-
-	// populate references to file descriptor dependencies
-	files := map[string]*FileDescriptor{}
-	for _, f := range deps {
-		files[f.proto.GetName()] = f
-	}
-	ret.deps = make([]*FileDescriptor, len(fd.GetDependency()))
-	for i, d := range fd.GetDependency() {
-		ret.deps[i] = files[d]
-		if ret.deps[i] == nil {
-			return nil, fmt.Errorf("Given dependencies did not include %q", d)
-		}
-	}
-	ret.publicDeps = make([]*FileDescriptor, len(fd.GetPublicDependency()))
-	for i, pd := range fd.GetPublicDependency() {
-		ret.publicDeps[i] = ret.deps[pd]
-	}
-	ret.weakDeps = make([]*FileDescriptor, len(fd.GetWeakDependency()))
-	for i, wd := range fd.GetWeakDependency() {
-		ret.weakDeps[i] = ret.deps[wd]
-	}
-	ret.isProto3 = fd.GetSyntax() == "proto3"
-
-	// populate all tables of child descriptors
-	for _, m := range fd.GetMessageType() {
-		md, n := createMessageDescriptor(ret, ret, pkg, m, ret.symbols)
-		ret.symbols[n] = md
-		ret.messages = append(ret.messages, md)
-	}
-	for _, e := range fd.GetEnumType() {
-		ed, n := createEnumDescriptor(ret, ret, pkg, e, ret.symbols)
-		ret.symbols[n] = ed
-		ret.enums = append(ret.enums, ed)
-	}
-	for _, ex := range fd.GetExtension() {
-		exd, n := createFieldDescriptor(ret, ret, pkg, ex)
-		ret.symbols[n] = exd
-		ret.extensions = append(ret.extensions, exd)
-	}
-	for _, s := range fd.GetService() {
-		sd, n := createServiceDescriptor(ret, pkg, s, ret.symbols)
-		ret.symbols[n] = sd
-		ret.services = append(ret.services, sd)
-	}
-
-	ret.sourceInfo = descinternal.CreateSourceInfoMap(fd)
-	ret.sourceInfoRecomputeFunc = ret.recomputeSourceInfo
-
-	// now we can resolve all type references and source code info
-	scopes := []scope{fileScope(ret)}
-	path := make([]int32, 1, 8)
-	path[0] = descinternal.File_messagesTag
-	for i, md := range ret.messages {
-		if err := md.resolve(append(path, int32(i)), scopes); err != nil {
-			return nil, err
-		}
-	}
-	path[0] = descinternal.File_enumsTag
-	for i, ed := range ret.enums {
-		ed.resolve(append(path, int32(i)))
-	}
-	path[0] = descinternal.File_extensionsTag
-	for i, exd := range ret.extensions {
-		if err := exd.resolve(append(path, int32(i)), scopes); err != nil {
-			return nil, err
-		}
-	}
-	path[0] = descinternal.File_servicesTag
-	for i, sd := range ret.services {
-		if err := sd.resolve(append(path, int32(i)), scopes); err != nil {
-			return nil, err
-		}
-	}
-
-	return ret, nil
-}
-
-// CreateFileDescriptors constructs a set of descriptors, one for each of the
-// given descriptor protos. The given set of descriptor protos must include all
-// transitive dependencies for every file.
-func CreateFileDescriptors(fds []*dpb.FileDescriptorProto) (map[string]*FileDescriptor, error) {
-	if len(fds) == 0 {
-		return nil, nil
-	}
-	files := map[string]*dpb.FileDescriptorProto{}
-	resolved := map[string]*FileDescriptor{}
-	var name string
-	for _, fd := range fds {
-		name = fd.GetName()
-		files[name] = fd
-	}
-	for _, fd := range fds {
-		_, err := createFromSet(fd.GetName(), nil, files, resolved)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return resolved, nil
-}
-
-// ToFileDescriptorSet creates a FileDescriptorSet proto that contains all of the given
-// file descriptors and their transitive dependencies. The files are topologically sorted
-// so that a file will always appear after its dependencies.
-func ToFileDescriptorSet(fds ...*FileDescriptor) *dpb.FileDescriptorSet {
-	var fdps []*dpb.FileDescriptorProto
-	addAllFiles(fds, &fdps, map[string]struct{}{})
-	return &dpb.FileDescriptorSet{File: fdps}
-}
-
-func addAllFiles(src []*FileDescriptor, results *[]*dpb.FileDescriptorProto, seen map[string]struct{}) {
-	for _, fd := range src {
-		if _, ok := seen[fd.GetName()]; ok {
-			continue
-		}
-		seen[fd.GetName()] = struct{}{}
-		addAllFiles(fd.GetDependencies(), results, seen)
-		*results = append(*results, fd.AsFileDescriptorProto())
-	}
-}
-
-// CreateFileDescriptorFromSet creates a descriptor from the given file descriptor set. The
-// set's *last* file will be the returned descriptor. The set's remaining files must comprise
-// the full set of transitive dependencies of that last file. This is the same format and
-// order used by protoc when emitting a FileDescriptorSet file with an invocation like so:
-//    protoc --descriptor_set_out=./test.protoset --include_imports -I. test.proto
-func CreateFileDescriptorFromSet(fds *dpb.FileDescriptorSet) (*FileDescriptor, error) {
-	files := fds.GetFile()
-	if len(files) == 0 {
-		return nil, errors.New("file descriptor set is empty")
-	}
-	resolved, err := CreateFileDescriptors(files)
-	if err != nil {
-		return nil, err
-	}
-	lastFilename := files[len(files)-1].GetName()
-	return resolved[lastFilename], nil
-}
-
-// createFromSet creates a descriptor for the given filename. It recursively
-// creates descriptors for the given file's dependencies.
-func createFromSet(filename string, seen []string, files map[string]*dpb.FileDescriptorProto, resolved map[string]*FileDescriptor) (*FileDescriptor, error) {
-	for _, s := range seen {
-		if filename == s {
-			return nil, fmt.Errorf("cycle in imports: %s", strings.Join(append(seen, filename), " -> "))
-		}
-	}
-	seen = append(seen, filename)
-
-	if d, ok := resolved[filename]; ok {
-		return d, nil
-	}
-	fdp := files[filename]
-	if fdp == nil {
-		return nil, fmt.Errorf("file descriptor set missing a dependency: %s", filename)
-	}
-	deps := make([]*FileDescriptor, len(fdp.GetDependency()))
-	for i, depName := range fdp.GetDependency() {
-		if dep, err := createFromSet(depName, seen, files, resolved); err != nil {
-			return nil, err
-		} else {
-			deps[i] = dep
-		}
-	}
-	d, err := CreateFileDescriptor(fdp, deps...)
-	if err != nil {
-		return nil, err
-	}
-	resolved[filename] = d
-	return d, nil
-}
-
 func (fd *FileDescriptor) recomputeSourceInfo() {
-	descinternal.PopulateSourceInfoMap(fd.proto, fd.sourceInfo)
+	internal.PopulateSourceInfoMap(fd.proto, fd.sourceInfo)
 }
 
 func (fd *FileDescriptor) registerField(field *FieldDescriptor) {
@@ -496,30 +310,30 @@ func createMessageDescriptor(fd *FileDescriptor, parent Descriptor, enclosing st
 
 func (md *MessageDescriptor) resolve(path []int32, scopes []scope) error {
 	md.sourceInfoPath = append([]int32(nil), path...) // defensive copy
-	path = append(path, descinternal.Message_nestedMessagesTag)
+	path = append(path, internal.Message_nestedMessagesTag)
 	scopes = append(scopes, messageScope(md))
 	for i, nmd := range md.nested {
 		if err := nmd.resolve(append(path, int32(i)), scopes); err != nil {
 			return err
 		}
 	}
-	path[len(path)-1] = descinternal.Message_enumsTag
+	path[len(path)-1] = internal.Message_enumsTag
 	for i, ed := range md.enums {
 		ed.resolve(append(path, int32(i)))
 	}
-	path[len(path)-1] = descinternal.Message_fieldsTag
+	path[len(path)-1] = internal.Message_fieldsTag
 	for i, fld := range md.fields {
 		if err := fld.resolve(append(path, int32(i)), scopes); err != nil {
 			return err
 		}
 	}
-	path[len(path)-1] = descinternal.Message_extensionsTag
+	path[len(path)-1] = internal.Message_extensionsTag
 	for i, exd := range md.extensions {
 		if err := exd.resolve(append(path, int32(i)), scopes); err != nil {
 			return err
 		}
 	}
-	path[len(path)-1] = descinternal.Message_oneOfsTag
+	path[len(path)-1] = internal.Message_oneOfsTag
 	for i, od := range md.oneOfs {
 		od.resolve(append(path, int32(i)))
 	}
@@ -1253,7 +1067,7 @@ func (sv sortedValues) Swap(i, j int) {
 
 func (ed *EnumDescriptor) resolve(path []int32) {
 	ed.sourceInfoPath = append([]int32(nil), path...) // defensive copy
-	path = append(path, descinternal.Enum_valuesTag)
+	path = append(path, internal.Enum_valuesTag)
 	for i, evd := range ed.values {
 		evd.resolve(append(path, int32(i)))
 	}
@@ -1463,7 +1277,7 @@ func createServiceDescriptor(fd *FileDescriptor, enclosing string, sd *dpb.Servi
 
 func (sd *ServiceDescriptor) resolve(path []int32, scopes []scope) error {
 	sd.sourceInfoPath = append([]int32(nil), path...) // defensive copy
-	path = append(path, descinternal.Service_methodsTag)
+	path = append(path, internal.Service_methodsTag)
 	for i, md := range sd.methods {
 		if err := md.resolve(append(path, int32(i)), scopes); err != nil {
 			return err
@@ -1776,7 +1590,7 @@ func fileScope(fd *FileDescriptor) scope {
 	// we search symbols in this file, but also symbols in other files that have
 	// the same package as this file or a "parent" package (in protobuf,
 	// packages are a hierarchy like C++ namespaces)
-	prefixes := descinternal.CreatePrefixList(fd.proto.GetPackage())
+	prefixes := internal.CreatePrefixList(fd.proto.GetPackage())
 	return func(name string) Descriptor {
 		for _, prefix := range prefixes {
 			n := merge(prefix, name)
@@ -1849,300 +1663,4 @@ func merge(a, b string) string {
 	} else {
 		return a + "." + b
 	}
-}
-
-var (
-	cacheMu       sync.RWMutex
-	filesCache    = map[string]*FileDescriptor{}
-	messagesCache = map[string]*MessageDescriptor{}
-	enumCache     = map[reflect.Type]*EnumDescriptor{}
-)
-
-// LoadFileDescriptor creates a file descriptor using the bytes returned by
-// proto.FileDescriptor. Descriptors are cached so that they do not need to be
-// re-processed if the same file is fetched again later.
-func LoadFileDescriptor(file string) (*FileDescriptor, error) {
-	f := getFileFromCache(file)
-	if f != nil {
-		return f, nil
-	}
-	cacheMu.Lock()
-	defer cacheMu.Unlock()
-	return loadFileDescriptorLocked(file)
-}
-
-func loadFileDescriptorLocked(file string) (*FileDescriptor, error) {
-	f := filesCache[file]
-	if f != nil {
-		return f, nil
-	}
-	fd, err := internal.LoadFileDescriptor(file)
-	if err != nil {
-		return nil, err
-	}
-
-	f, err = toFileDescriptorLocked(fd)
-	if err != nil {
-		return nil, err
-	}
-	putCacheLocked(file, f)
-	return f, nil
-}
-
-func toFileDescriptorLocked(fd *dpb.FileDescriptorProto) (*FileDescriptor, error) {
-	deps := make([]*FileDescriptor, len(fd.GetDependency()))
-	for i, dep := range fd.GetDependency() {
-		var err error
-		deps[i], err = loadFileDescriptorLocked(dep)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return CreateFileDescriptor(fd, deps...)
-}
-
-func getFileFromCache(file string) *FileDescriptor {
-	cacheMu.RLock()
-	defer cacheMu.RUnlock()
-	return filesCache[file]
-}
-
-func putCacheLocked(filename string, fd *FileDescriptor) {
-	filesCache[filename] = fd
-	putMessageCacheLocked(fd.messages)
-}
-
-func putMessageCacheLocked(mds []*MessageDescriptor) {
-	for _, md := range mds {
-		messagesCache[md.fqn] = md
-		putMessageCacheLocked(md.nested)
-	}
-}
-
-// interface implemented by generated messages, which all have a Descriptor() method in
-// addition to the methods of proto.Message
-type protoMessage interface {
-	proto.Message
-	Descriptor() ([]byte, []int)
-}
-
-// LoadMessageDescriptor loads descriptor using the encoded descriptor proto returned by
-// Message.Descriptor() for the given message type. If the given type is not recognized,
-// then a nil descriptor is returned.
-func LoadMessageDescriptor(message string) (*MessageDescriptor, error) {
-	m := getMessageFromCache(message)
-	if m != nil {
-		return m, nil
-	}
-
-	pt := proto.MessageType(message)
-	if pt == nil {
-		return nil, nil
-	}
-	msg, err := messageFromType(pt)
-	if err != nil {
-		return nil, err
-	}
-
-	cacheMu.Lock()
-	defer cacheMu.Unlock()
-	return loadMessageDescriptorForTypeLocked(message, msg)
-}
-
-// LoadMessageDescriptorForType loads descriptor using the encoded descriptor proto returned
-// by message.Descriptor() for the given message type. If the given type is not recognized,
-// then a nil descriptor is returned.
-func LoadMessageDescriptorForType(messageType reflect.Type) (*MessageDescriptor, error) {
-	m, err := messageFromType(messageType)
-	if err != nil {
-		return nil, err
-	}
-	return LoadMessageDescriptorForMessage(m)
-}
-
-// LoadMessageDescriptorForMessage loads descriptor using the encoded descriptor proto
-// returned by message.Descriptor(). If the given type is not recognized, then a nil
-// descriptor is returned.
-func LoadMessageDescriptorForMessage(message proto.Message) (*MessageDescriptor, error) {
-	// efficiently handle dynamic messages
-	type descriptorable interface {
-		GetMessageDescriptor() *MessageDescriptor
-	}
-	if d, ok := message.(descriptorable); ok {
-		return d.GetMessageDescriptor(), nil
-	}
-
-	name := proto.MessageName(message)
-	if name == "" {
-		return nil, nil
-	}
-	m := getMessageFromCache(name)
-	if m != nil {
-		return m, nil
-	}
-
-	cacheMu.Lock()
-	defer cacheMu.Unlock()
-	return loadMessageDescriptorForTypeLocked(name, message.(protoMessage))
-}
-
-func messageFromType(mt reflect.Type) (protoMessage, error) {
-	if mt.Kind() != reflect.Ptr {
-		mt = reflect.PtrTo(mt)
-	}
-	m, ok := reflect.Zero(mt).Interface().(protoMessage)
-	if !ok {
-		return nil, fmt.Errorf("failed to create message from type: %v", mt)
-	}
-	return m, nil
-}
-
-func loadMessageDescriptorForTypeLocked(name string, message protoMessage) (*MessageDescriptor, error) {
-	m := messagesCache[name]
-	if m != nil {
-		return m, nil
-	}
-
-	fdb, _ := message.Descriptor()
-	fd, err := internal.DecodeFileDescriptor(name, fdb)
-	if err != nil {
-		return nil, err
-	}
-
-	f, err := toFileDescriptorLocked(fd)
-	if err != nil {
-		return nil, err
-	}
-	putCacheLocked(fd.GetName(), f)
-	return f.FindSymbol(name).(*MessageDescriptor), nil
-}
-
-func getMessageFromCache(message string) *MessageDescriptor {
-	cacheMu.RLock()
-	defer cacheMu.RUnlock()
-	return messagesCache[message]
-}
-
-// interface implemented by all generated enums
-type protoEnum interface {
-	EnumDescriptor() ([]byte, []int)
-}
-
-// NB: There is no LoadEnumDescriptor that takes a fully-qualified enum name because
-// it is not useful since protoc-gen-go does not expose the name anywhere in generated
-// code or register it in a way that is it accessible for reflection code. This also
-// means we have to cache enum descriptors differently -- we can only cache them as
-// they are requested, as opposed to caching all enum types whenever a file descriptor
-// is cached. This is because we need to know the generated type of the enums, and we
-// don't know that at the time of caching file descriptors.
-
-// LoadEnumDescriptorForType loads descriptor using the encoded descriptor proto returned
-// by enum.EnumDescriptor() for the given enum type.
-func LoadEnumDescriptorForType(enumType reflect.Type) (*EnumDescriptor, error) {
-	// we cache descriptors using non-pointer type
-	if enumType.Kind() == reflect.Ptr {
-		enumType = enumType.Elem()
-	}
-	e := getEnumFromCache(enumType)
-	if e != nil {
-		return e, nil
-	}
-	enum, err := enumFromType(enumType)
-	if err != nil {
-		return nil, err
-	}
-
-	cacheMu.Lock()
-	defer cacheMu.Unlock()
-	return loadEnumDescriptorForTypeLocked(enumType, enum)
-}
-
-// LoadEnumDescriptorForEnum loads descriptor using the encoded descriptor proto
-// returned by enum.EnumDescriptor().
-func LoadEnumDescriptorForEnum(enum protoEnum) (*EnumDescriptor, error) {
-	et := reflect.TypeOf(enum)
-	// we cache descriptors using non-pointer type
-	if et.Kind() == reflect.Ptr {
-		et = et.Elem()
-		enum = reflect.Zero(et).Interface().(protoEnum)
-	}
-	e := getEnumFromCache(et)
-	if e != nil {
-		return e, nil
-	}
-
-	cacheMu.Lock()
-	defer cacheMu.Unlock()
-	return loadEnumDescriptorForTypeLocked(et, enum)
-}
-
-func enumFromType(et reflect.Type) (protoEnum, error) {
-	if et.Kind() != reflect.Int32 {
-		et = reflect.PtrTo(et)
-	}
-	e, ok := reflect.Zero(et).Interface().(protoEnum)
-	if !ok {
-		return nil, fmt.Errorf("failed to create enum from type: %v", et)
-	}
-	return e, nil
-}
-
-func loadEnumDescriptorForTypeLocked(et reflect.Type, enum protoEnum) (*EnumDescriptor, error) {
-	e := enumCache[et]
-	if e != nil {
-		return e, nil
-	}
-
-	fdb, path := enum.EnumDescriptor()
-	name := fmt.Sprintf("%v", et)
-	fd, err := internal.DecodeFileDescriptor(name, fdb)
-	if err != nil {
-		return nil, err
-	}
-	// see if we already have cached "rich" descriptor
-	f, ok := filesCache[fd.GetName()]
-	if !ok {
-		f, err = toFileDescriptorLocked(fd)
-		if err != nil {
-			return nil, err
-		}
-		putCacheLocked(fd.GetName(), f)
-	}
-
-	ed := findEnum(f, path)
-	enumCache[et] = ed
-	return ed, nil
-}
-
-func getEnumFromCache(et reflect.Type) *EnumDescriptor {
-	cacheMu.RLock()
-	defer cacheMu.RUnlock()
-	return enumCache[et]
-}
-
-func findEnum(fd *FileDescriptor, path []int) *EnumDescriptor {
-	if len(path) == 1 {
-		return fd.GetEnumTypes()[path[0]]
-	}
-	md := fd.GetMessageTypes()[path[0]]
-	for _, i := range path[1 : len(path)-1] {
-		md = md.GetNestedMessageTypes()[i]
-	}
-	return md.GetNestedEnumTypes()[path[len(path)-1]]
-}
-
-// LoadFieldDescriptorForExtension loads the field descriptor that corresponds to the given
-// extension description.
-func LoadFieldDescriptorForExtension(ext *proto.ExtensionDesc) (*FieldDescriptor, error) {
-	file, err := LoadFileDescriptor(ext.Filename)
-	if err != nil {
-		return nil, err
-	}
-	field, ok := file.FindSymbol(ext.Name).(*FieldDescriptor)
-	// make sure descriptor agrees with attributes of the ExtensionDesc
-	if !ok || !field.IsExtension() || field.GetOwner().GetFullyQualifiedName() != proto.MessageName(ext.ExtendedType) ||
-		field.GetNumber() != ext.Field {
-		return nil, fmt.Errorf("File descriptor contained unexpected object with name %s:", ext.Name)
-	}
-	return field, nil
 }
